src/data_pipeline/realTime_data/functionapp_httpTrigger_for_bronze_to_silver_DFpipline/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-mgmt-resource
azure-mgmt-datafactory
azure-identity